backend/sop_agent/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
PyJWT==2.8.0
google-generativeai==0.3.2
psycopg2-binary==2.9.9
pgvector==0.2.4
sentence-transformers==2.2.2
numpy==1.24.4
flask_cors
